[Silver I] Title: 오르막 수, Time: 0 ms, Memory: 2060 KB -BaekjoonHub

+++ 백준/Silver/11057. 오르막 수/오르막 수.cc
@@ -3,7 +3,7 @@
 using namespace std;
 
 int n;
-long long dp[1001][10];		// dp[i][j] : i자리 오르막 수 중 j로 끝나는 수의 개수 
+int dp[1001][10];		// dp[i][j] : i자리 오르막 수 중 j로 끝나는 수의 개수 
 const int MOD = 10007;
 
 int solution()
